requirements.txt
pbr>=0.5.21,<1.0
anyjson
eventlet>=0.9.12
jsonpath
Paste
PasteDeploy
iso8601>=0.1.4
python-heatclient>=0.2.4
jsonschema==2.0.0
netaddr

oslo.config
deep
murano-common>=0.2.2
PyYAML>=3.1.0


# temporary until released
https://github.com/EkaterinaFedorova/python-metadataclient/archive/master.tar.gz#egg=metadataclient
https://github.com/stackforge/murano-repository/archive/master.tar.gz#egg=muranorepository
